Integrate CourseTimelineMethods.java methods into User
